Restructure project
From what I can tell the structure proposed in the book can not
easily be extended in the way it is created.

From what I gather, if you had a library in your project that you
wanted to make its own crate, the way to go would be to move it to
its own directory with its own Cargo.toml
(or cargo new --lib my_library) and include that in the main
executable's Cargo.toml as a dependency with an explicit path.